+(*)Rescale variables with homogenize_forcing
  Add optional tmp_scale arguments to the various homogenize_field routines and
use them in numerous calls in homogenize_forcing and homogenize_mech_forcing, so
that the answers will pass the dimensional rescaling tests for large rescaling
factors when HOMOGENIZE_FORCINGS = True.  Among other changes is the addition of
a verticalGrid_type argument to homogenize_forcing.  All answers are bitwise
identical in the existing MOM6-examples test suite, but it will change answers
in other cases with dimensional rescaling active and homogenized forcing.