MOM_random: Set mask with bits rather than integer
The MOM_random generator used bit masks which were set with integer
values.  This is problematic for the sequence 0x8000 0000, because it
must be set with a value of -2**31.

In 4-byte integers, this is strictly not representable in Fortran, which
requires symmetric signed domains for its variables.  Since the upper
bound is 2**31 - 1, the lower bound must be -2**31 + 1, which is larger
than -2**31.  Any value assigned to the 0x80000000 bit sequence is
considered a noncompliant compiler extension.

The current implementation seems to resolve this by using a kind=8 value
(itself problematic, since 8-byte is not assured), but it still requires
assigning this value to a 4-byte integer which cannot (strictly)
represent the value.

This patch averts the whole issue by explicitly setting the bits, and
makes no reference to its integer value.  It leaves the compiler to
decide its interpretation.  And since the variable is only used in bit
operations, there is no ambiguity in behavior.

Note that GCC 9 does not support BOZ conversion from z'80000000' to int,
since it still expects BOZ literals to be within the bounds.  This is
why we use ibset() in place of a literal.  Later GCC versions do not
have this objection.