Remove unused variables
This patch removes any unused variables detected by -Wunused-variables
in GCC.

This may include variables which are referenced in commented "zombie
code".  In some cases, the variables were preserved in comments, but not
in all cases.  If the zombie code is restored, it should be possible to
reconstruct the original declaration.  (All known cases were simple
scalars, such as loop indices).

In one case, a variable was conditionally used within a preprocessor
`#ifdef __DO_SAFETY_CHECKS__` block.  This patch moves the declaration
into another `#ifdef` block, but perhaps it is preferable to remove the
block.

Certain documentation variables such as `mdl`, which typically contain
the function name, have been removed if unused.

The strongest motivation for this patch is to allow us to enable the
`-Wall` flag, which includes unused variables, and will strengthen our
ability to detect potential errors in the codebase.