FMS2: get_field_sizes with FMS1 emulation
This patch modifies the `get_field_sizes` function to emulate the
FMS1 behavior when the requested `sizes(:)` is larger than the field's
number of dimensions.

In FMS1, `field_sizes` would always expect a `sizes` array of at least
size 4, and in the format of (nx, ny, nz, nt), and would lean on
netCDF attributes like `cartesian_axes` when assigning the values.

FMS2 lacks similar functionality, due its more general approach.
Previously, this was emulated in a way favorable to certain situations,
but raised issues in others.

This patch attempts to resolve the issue by using the
`categorize_axes()` function's ability to identify axes, either by
attributes, presumed names, or the `unlimited` property.