+Refactored MOM_density_integrals
  Refactored MOM_density_integrals to use the newer calculate_density_1d() and
calculated_stanley_density_1d() interfaces to the equation of state routines,
and to thereby shift all related dimensional rescaling into MOM_EOS.F90.  Also
revised the comments describing the arguments to a number of the equation of
state routines to eliminate confusing options and clearly indicate the units of
each input and output variable.  As a part of this change, the units of the
rho_ref argument to calculate_stanley_density_1d were changed from [kg m-3] to
[R ~> kg m-3] to match the equivalent routine calculate_density_1d().  Because
this does not appear to have been used previously, this should not be a problem,
and answers will not change unless a dimensional consistency test is underway.
All answers are bitwise identical, but there is one minor change to the rescaled
units of one apparently unused optional argument.