+Add scale argument to set_up_ALE_sponge
  Added optional scaling arguments to the set_up_ALE_sponge routines, to allow
input fields to be rescaled before use.  This change is necessary to permit the
dimensional rescaling of temperature, salinity, and other tracers because of the
way that some versions repeatedly read new values from files as the runs
progress.  All answers are bitwise identical, but there are new optional
arguments to public interfaces.