Put CVMix convective viscosity into shear term
In both the legacy and non-legacy diabatic drivers, the viscosity
output from CVMix convection went to the `Kv_slow` term. However, this
term is only allocated if KPP is being used, leading to a segfault if
`USE_KPP = False`.

This may have been the intended purpose of `Kv_slow`, but its
contribution is already added to the shear term before convection is
calculated, so convection can't be accounted using `Kv_slow`. To
ensure we don't hit a segfault, and that the enhanced viscosity is
actually present, we change convection to go into `Kv_shear` in all
cases.