Replace findloc() with user-defined find_index
This patch replaces the instances of findloc() with a new function,
find_index.  They should be functionally equivalent.

findloc() is a F2008 intrinsic which returns the index of the first
instance of a value in an array.  Support is still spotty on some older
compilers in active use, so this function is the safer alternative.