Re-factored GFDL gitlab pipeline
The MOM6 pipeline commands were scripted using make in the file MOM6-examples/tools/MRS/Makefile. This was trying to encapsulate the commands used in the pipeline so they could also be used at the command line (for emulating the pipeline). However, it required emulating the temporary/transient work-spaces provided by gitlab. Also, the Makefile was impenetrable.

New approach:
- Create a persistent working space for each pipeline, just as you'd work interactively in a single working directory
- Use the same commands in .gitlab-ci.yml as we'd use interactively (these still use the tools/MRS Makefile.build and Makefile.run)
- The compute stage of tasks is now a bash script (.gitlab/mom6-ci-run-script.sh) which will work/can be submitted from your working directory
- The caching of results from each run (gnu/intel/pgi, symmetric/nonsymmetric/layout/etc) is stored locally in the same working directory so we don't have to look elsewhere for the results
- Using the "fetch" strategy allows later stages to startup in mere seconds (previously re-cloning the repo took minutes)

This does not shorten the turn around significantly but I believe it is a lot easy to follow and emulate.

Todo:
- split up the run stage to reduce it from 41+ minutes to ~13 minutes.