(*)Use visc%Kv_shear in calculate_CVMix_conv call
  Use visc%Kv_shear in the calculate_CVMix_conv call in layered_diabatic, to
avoid segmentation faults when USE_REGRIDDING=True and USE_CVMix_CONVECTION=True
but USE_KPP=False.  This bug was identified when evaluating the changes in
https://github.com/NOAA-GFDL/MOM6/pull/129, and a correction was requested, but
the correction was omitted from that PR.  Answers could change, but it is likely
that any cases that would change would previously have encountered a
segmentation fault, so I suspect that no such cases exist.