+Add DEBUG_FULL_COLUMN for debugging accelerations
  Added the runtime debugging option DEBUG_FULL_COLUMN to write out the
acceleration diagnostics for the entire water column when there are velocity
truncations, rather than just writing out the values for layers with large
velocities.  Also modified the write_accel code to reflect that the viscous
coupling coefficients are discretized at the interfaces, and to reflect the fact
that the stresses are optional arguments.  All answers are bitwise identical,
but there is a new debugging runtime argument, and the debugging output changes
when there are velocity truncations.