+Convert 16 vertvisc_type pointers to allocatables
  Converted 16 pointers in the vertvisc_type into allocatable arrays.  Also
added units describing some variables, and used whether the visc%Ray terms are
allocated to determine whether there are to be Rayleigh drag terms in the
tridiagonal viscosity solver, rather than using a duplicate call to get_param
for CHANNEL_DRAG to control this behavior.  All answers and output are bitwise
identical, but there are changes to the types of some of the elements of a
transparent type.