+Overload MOM_tracer_chksum to use tracer registry
  Overloaded MOM_tracer_chksum and MOM_tracer_chkinv with a simpler interface
that takes the tracer registry as an input argument, rather than requiring that
its elements be unpacked outside of the call.  This was done as an overload to
the existing interface to avoid breaking backward compatibility, but it seems
likely that in due course the older, more complicated interface can be
obsoleted.  All answers are bitwise identical, but there are new interfaces to
provide tracer debugging capabilities.