*Fix DOME2d_initialize_temperature_salinity bug
  Removed a line resetting temperatures to 0 for the ZSTAR and SIGMA vertical
coordinates in DOME2d_initialize_temperature_salinity, after they had been set
to the intended values.  github.com/mom-ocean/MOM6/issues/1560 highlights this
bug and can be closed after this commit is merged into the main branch of MOM6.
This changes the temperature fields (and ocean.stats files) for the
flow_downslopes/z and flow_downslopes/sigma test cases, but because temperature
does not influence density in these cases, the flows and salinities are
unchanged.