+*Use runtime parameters in DOME initialization
  Modified the DOME initialization to simplify some expressions and to use
run-time parameters, rather than hard-coded values, when initializing the DOME
test case.  By default, all expressions are mathematically equivalent, but there
are roundoff level changes in the topography and sponges due to the use of more
generally valid expressions with runtime parameters and the replacement of some
divisions for unit conversions by multiplication by a reciprocal.  Because the
DOME test cases are strongly nonlinear, these small changes cascade up to
macroscopic differences, but these are of comparable magnitude to the
differences between compilers.  There are 10 new runtime parameters that appear
in the MOM_parameter_doc.all files for the DOME test cases.