+Fix diagnostic conversions in code from dev/ncar
  Added missing unit conversion factors in two calls to register_diag_field and
modified the comments describing the units of some temperature and salinity
variables in code that was recently added to the recently main branch of MOM6
via dev/ncar, reflecting the rescaling of temperatures and salinities that is
now in the dev/gfdl branch.  Without this change, these diagnostics will fail
the dimensional consistency testing for temperature and salinity.  All answers
are bitwise identical, but there is a case change of the units of a variable
in the available_diags files.