Call set_initialized after query_initialized
  Call set_initialized for variables after they are initialized, when their lack
of initialization was detected via query_initialized.  This commit reproduces
the model's behavior prior to the removal of the code that set the initialized
flag within query_initialized in https://github.com/NOAA-GFDL/MOM6/pull/149, and
it completes the set of changes that was envisioned when set_initialized was
added in https://github.com/NOAA-GFDL/MOM6/pull/152.  All answers are bitwise
identical.