+Rescale sfc_state%SST and sfc_state%SSS
  Rescaled the SST and SSS element of the surface type, usually sfc_state%SST
and sfc_state%SSS, from units of [degC] and [ppt] to [degC ~> C] and [ppt ~> S],
as well as a handful of other temperature and salinity variables related to the
surface forcing (usually targets of restoring), and cancelled out a number of
common US%C_to_degC or US%S_to_ppt conversion factors.  Several unused variables
were also removed, and a missing allocate and restart registration were added
for the running-mean salinity in the (as yet unused) MOM_controlled_forcing
module.  All answers are bitwise identical, but there are changes to the
rescaled units of two elements of a public type.