(*)Fix set_up_ALE_sponge_field call T & S scaling
  Corrected the dimensional rescaling factors in two calls to
set_up_ALE_sponge_field for temperature and salinity for time-varying fields
being read in from an input file.  These had been given the inverse of the
correct values.  An optional scale argument was also added (with its default
value) in the call to set up ALE sponge velocities, for greater clarity of what
this call is doing.  This commit should address an issue noted by Kate Hedstrom
when evaluating the first draft of PR #1577 from dev/gfdl to main.  All answers
are bitwise identical in cases where dimensional rescaling of temperature and
salinity are not being applied, and answers with the rescaling of temperature
and salinity should now reproduce those without the rescaling.