Fatal error if a tracer package is registered twice
  Made the WARNING error messages that were issued if a tracer package
registration routine is called with an associated control structure into FATAL
errors, as suggested by github.com/mom-ocean/MOM6/issues/1444.  A tracer package
could be used twice, but it would require the use of separate control structures
for each instance.  This change will bring down the model in a case that
probably should not be running.  All solutions are bitwise identical in any
cases that run, and all MOM6-examples test cases work exactly as before.