Autoconf: Logging for dependency builds
This patch adds the `REPORT_ERROR_LOGS` variable to the .testing and
ac/deps makefiles.  When set to true, a failed FMS build will also echo
the contents of `config.log`, in order to diagnose remote CI builds.

The variable is disabled on default, to reduce the amount of output at
command line.  It has been added to the GitHub Actions CI, and is
propagated through `.testing/Makefile`.

Although not used in the MOM6 builds, it could be extended to them if it
proves useful.  Currently, these Makefiles always echo `config.log`
after a failed build.