CI: Replace GitHub Actions MacOS compiler env
This patch fixes the gcc compiler environment variables on their MacOS
systems.

GitHub recently made some GCC compiler changes on their MacOS machines;
specifically, `gcc-11` and `gfortran-11` appear to no longer be
available.  Previously, we were unable to use the default gcc and had to
use these executables.

On this newer version, we can use the default system gcc, and the `CC`
environment variable can be left unset.  We still need `FC` to point to
`gfortran`, since it otherwise tries to build with `f77`, which does not
exist.