+(*)Set ustar with wind_forcing_2gyre
  Set the friction velocity, ustar, when the wind stresses are specified via
wind_forcing_2gyre and wind_forcing_2gyre.  This was done by adding the new
subroutine stresses_to_ustar to start to standardize the calculation of the
friction velocities from the wind stresses.  Neverworld_wind_forcing and
wind_forcing_gyres were already using the same expressions for ustar as are in
stresses_to_ustar, so the new routine is now used in these cases.  In other
routines that set ustar, the answers differ at roundoff due to the difference
between division by a reference density and multiplication by its reciprocal, so
using the new subroutine in those cases should be wrapped in an answer_date
flag, which is being left for a later commit.  This is being added now as a
result of a new commit that uses ustar as a part of the specification of
viscosities even in idealized cases.  All answers in the MOM6-examples test
suite are bitwise identical.